+++ CHANGELOG.md
@@ -35,6 +35,7 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 
 - Fix deprecation message for `Modifier::uri()` method [166](https://github.com/thephpleague/uri-src/pull/166) by [meyerbaptiste](https://github.com/meyerbaptiste) 
 - Fix missing SensitiveParameter usage on `Modifier::withUserInfo` method
+- Fix Host resolution using the new `HostRecord` class
 
 ### Deprecated